Finished the confusions !
when using two items with the same name and the same type (bone, or animation), some strange things appeared. Its is no longer doing that way since the references are based on the internal ID instead of the name. So now a new bigger problem appears, see #b1e3e3b in the SFDynamicObject repositery.